Move out MiroIdentifiers into a separate trait
